requirements.txt
torch
torchvision
ultralytics
timm>=0.9.0
opencv-python
pillow
numpy
pandas
scikit-learn
albumentations
matplotlib
seaborn
grad-cam
pyyaml
tqdm
wandb
pycocotools
pydicom
pytest
streamlit
